fix: reach zero pyright errors in the CI gate environment
The 3.10 CI runner resolves tomllib and sklearn/numpy differently from
a 3.12 local venv: make the tomllib fallback statically evaluable via
sys.version_info, and coerce the sklearn score results that the 3.10
resolution infers as possibly-ndarray. Verified 0 errors in both a
CI-parity 3.10 venv (2,850 warnings — the recorded gate baseline) and
the local 3.12 venv.

+++ limen/log/_permutation_confusion_metrics.py
@@ -189,7 +189,7 @@ def _permutation_confusion_metrics(self: Any,
 
         sp = sqrt(sp_num / sp_den) if sp_num > 0 else np.nan
 
-        return (ma - mb) / sp if sp and not np.isnan(sp) else np.nan
+        return float((ma - mb) / sp) if sp and not np.isnan(sp) else np.nan
 
     def _ks(a: np.ndarray, b: np.ndarray) -> float:
